feat: Ladder 값 객체 생성, 연속으로 건널 수 있는 다리 생성되는 버그 수정
